Move default teams to the pokemon module
