+++ pokemon_helpers/pokemon.py
@@ -123,3 +123,18 @@ class Pokemon:
         if key == "baseStats":
             key = "base_stats"
         return self.__getattribute__(key)
+
+
+def default_team_spinda():
+    """Generate a Spinda for these players."""
+    return [Pokemon(name="spinda", moves=["return", "shadowball", "tackle", "icebeam"])]
+
+
+def default_team_floatzel():
+    """Generate a FLoatzel for the player."""
+    return [Pokemon(name="floatzel", moves=["watergun", "tackle", "liquidation", "icebeam"])]
+
+
+def default_team_ivysaur():
+    """Generate an Ivysaur for these players."""
+    return [Pokemon(name="ivysaur", moves=["seedbomb", "tackle", "icebeam"])]
